% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_date.R
\name{add_date_range}
\alias{add_date_range}
\title{Add a date_range attribute to a dataset}
\usage{
add_date_range(dataset)
}
\arguments{
\item{dataset}{a dataset loaded with dv.loader}
}
\description{
This function will add an attribute \code{date_range} to a dataset.
The source for this dates is the \code{mtime} (POSIXct) entry from the \code{meta} attribute as provided by \code{dv.loader}.
As different domains are contained as a list in a single dataset, this attribute will contain the minimum
and the maximum modification date from all the domains in the dataset.

See \code{vignette("customizing_application")}
}
\details{
In case any of the entries in dataset does not have the expected attribute, it will show a warning.
}
