% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{The application User-Interface using an HTML template}
\usage{
app_ui(id)
}
\arguments{
\item{id}{This parameter can be either an string id if the app_ui is being used as a module or
a request if the app_ui is being used as an standalone application created in golem then it becomes
a Internal parameter for \code{{shiny}}
DO NOT REMOVE.}
}
\description{
The application User-Interface using an HTML template
}
\keyword{internal}
