% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{binary_deserialize_filter_data_C}
\alias{binary_deserialize_filter_data_C}
\title{Serialize filter data (binary form)}
\usage{
binary_deserialize_filter_data_C(x)
}
\description{
Serialize filter data (binary form)
}
