% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{build_secure_arguments}
\alias{build_secure_arguments}
\title{Securize an application using an Azure}
\usage{
build_secure_arguments(azure_options, app_ui, app_server)
}
\arguments{
\item{azure_options}{A named vector/list containing the options for the Azure Authentication:
Required fields: redirect, resource, tenant, app, version, password.
See AzureAuth::build_authorization_uri for these. If NULL then application will not be secure.}

\item{app_ui, app_server}{the shiny ui and server functions}
}
\value{
A list with the entries \code{ui_func} and \code{srv_func} and \code{options} that contains the securized version of the app.
}
\description{
Securize an application using an Azure
}
