% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{combine_filter_info}
\alias{combine_filter_info}
\title{Combine subject- and dataset-level filter results into one filter_info}
\usage{
combine_filter_info(filter_info)
}
\arguments{
\item{filter_info}{\code{list(result = list(subject =, dataset =), error_list =)}, as returned by \code{\link[=get_filter_info]{get_filter_info()}}.}
}
\value{
\code{list(result = list(subjects =, filter_info =), error_list =)}.
}
\description{
Can be expanded to combine arbitrary filters, not necessarily just these two.
}
