% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{flatten_srv_module_list}
\alias{flatten_srv_module_list}
\title{Flatten a List of Server Modules}
\usage{
flatten_srv_module_list(x)
}
\arguments{
\item{x}{A list of server modules, which may contain nested lists of server collections.}
}
\value{
A flattened list containing the \code{server} and \code{module_id} of each module.
}
\description{
This function recursively flattens a nested list structure of server modules into a single-level list.
It retains only the \code{server} and \code{module_id} for each entry that is not a server collection.
}
\keyword{internal}
