% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_date.R
\name{get_date_range}
\alias{get_date_range}
\title{Return a range of dates}
\usage{
get_date_range(x)
}
\arguments{
\item{x}{A list of date objects}
}
\value{
A character vector with two entries, the first and last date in the input vector.
}
\description{
Calculates a range of dates from a list of POSIXct objects. If any of the entries is NULL it will show
a warning indicating which of the entries
was NULL. If any of the entries is not a date object it will throw an error.
}
\keyword{internal}
