% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-label.R
\name{get_lbl}
\alias{get_lbl}
\title{Get a single label from a dataframe}
\usage{
get_lbl(df, var)
}
\arguments{
\item{df}{a dataframe.}

\item{var}{from which column do we want the label}
}
\value{
A string with the label name
}
\description{
It will try to read the label attribute from one column.
}
\keyword{internal}
