% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-label.R
\name{get_lbl_robust}
\alias{get_lbl_robust}
\title{Robust getter for dataframe labels}
\usage{
get_lbl_robust(df, var)
}
\arguments{
\item{df}{a dataframe.}

\item{var}{a column of the dataframe}
}
\value{
The label attribute or the name of the column if the label attribute is NULL.
}
\description{
It returns the label of the column of a dataframe if it is not NULL. Otherwise it returns the name of the column.
}
\keyword{internal}
