% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-label.R
\name{get_lbls}
\alias{get_lbls}
\alias{get_lbls_robust}
\title{Get all labels from a dataframe}
\usage{
get_lbls(df)

get_lbls_robust(df)
}
\arguments{
\item{df}{a dataframe.}
}
\value{
A named list in which name is the name of the df column and value is the label of the column.
Non-labelled columns will return NULL. In the robust version NULL is replaced by the name of the column.
}
\description{
It will try to read the label attribute from each of the columns.
}
\section{Functions}{
\itemize{
\item \code{get_lbls_robust()}: robust

}}
\keyword{internal}
