% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check.R
\name{has_all_items_named}
\alias{has_all_items_named}
\title{tests if all elements of a list are named}
\usage{
has_all_items_named(x)
}
\arguments{
\item{x}{a list to be tested}
}
\description{
Is true if all elements have a name or the is of length(0).
}
\keyword{internal}
