% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{include_css}
\alias{include_css}
\title{Include CSS files from path \verb{/inst/css/}}
\usage{
include_css(pattern = "*")
}
\arguments{
\item{pattern}{Default search for file pattern}
}
\description{
Imports custom CSS files
}
\keyword{internal}
