% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{include_js}
\alias{include_js}
\title{Include JS files from path \verb{/inst/js/}}
\usage{
include_js(pattern = "*")
}
\arguments{
\item{pattern}{Default search for file pattern}
}
\description{
Imports custom JS files
}
\keyword{internal}
