% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_logging_main.R
\name{log_with_logging_handlers}
\alias{log_with_logging_handlers}
\title{globalCallinHandlers do not capture the output of UI and Server for some reason therefore
specific with handlers are applied to them
Because there is no easy way of passing those handlers we use an option to pass those.}
\usage{
log_with_logging_handlers(fnc)
}
\description{
globalCallinHandlers do not capture the output of UI and Server for some reason therefore
specific with handlers are applied to them
Because there is no easy way of passing those handlers we use an option to pass those.
}
\keyword{internal}
