% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchers_function.R
\name{mm_dispatch}
\alias{mm_dispatch}
\title{Functional dispatcher helper functions}
\usage{
mm_dispatch(from, selection = character(0))
}
\arguments{
\item{from}{a string indicating what structure do we want to access inside module manager}

\item{selection}{an optional subsetting within the selected data structure}
}
\description{
Dispatch is a helper function that allows accessing data and utility functions inside module manager in
a dynamic way.
}
\details{
The parameter \code{from} can take the following values:
\itemize{
\item unfiltered_dataset
\item filtered_dataset
\item url_parameters
\item module_output
\item utils
}
}
\examples{

mm_dispatch("unfiltered_dataset", "adsl")
mm_dispatch("utils", "switch2")
}
\seealso{
mm_resolve_dispatcher
}
