% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchers_function.R
\name{mm_resolve_dispatcher}
\alias{mm_resolve_dispatcher}
\title{Dispatcher resolver}
\usage{
mm_resolve_dispatcher(dispatcher, afmm, flatten = FALSE)
}
\arguments{
\item{dispatcher}{A list of class "mm_dispatcher" to be resolved. If the object is any of other class it
is returned unmodified.}

\item{afmm}{a list of arguments from module manager}

\item{flatten}{for dispatcher where selection length is 1 and disppatcher is of class "mm_dispatcher"
return the first element instead of a one element list. By default, it will try to flatten them.
If selection is greater than 1 flatten must be FALSE, otherwise an error will be thrown.}
}
\description{
Dispatcher resolver
}
\seealso{
mm_dispatch
}
