% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_module.R
\name{mod_afmm_export}
\alias{mod_afmm_export}
\title{A testing module that exports the whole afmm object}
\usage{
mod_afmm_export(mod_id)
}
\arguments{
\item{mod_id}{shiny module ID}
}
\value{
A module list with ui, server, and module_id
}
\description{
This module is used for testing purposes to access the full afmm object
via shinytest2's get_value(export = "module_id-afmm").
}
\keyword{internal}
