% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app_info.R
\name{mod_app_info}
\alias{mod_app_info}
\title{A module to inspect the app information}
\usage{
mod_app_info(mod_id = ".._app_info..")
}
\arguments{
\item{mod_id}{Shiny module id. Default: \code{app_info}}
}
\description{
A module to inspect the app information
}
