% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_module.R
\name{mod_dataset_labels}
\alias{mod_dataset_labels}
\title{A simple module that lists the column labels of all used datasets}
\usage{
mod_dataset_labels(dataset_names, module_id)
}
\arguments{
\item{module_id}{shiny module ID}
}
\description{
This simple module is used for demonstration purposes in documentation
}
\keyword{internal}
