% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_module.R
\name{mod_export_dataset_name}
\alias{mod_export_dataset_name}
\title{A module that exports the current dataset list's name, for testing the export feature}
\usage{
mod_export_dataset_name(mod_id)
}
\arguments{
\item{mod_id}{shiny module ID}
}
\description{
A module that exports the current dataset list's name, for testing the export feature
}
\keyword{internal}
