% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_module.R
\name{multi_simple_UI}
\alias{multi_simple_UI}
\alias{multi_simple_server}
\alias{mod_multi_simple}
\title{A simple module that counts the number of rows for all datasets loaded in the application}
\usage{
multi_simple_UI(id)

multi_simple_server(id, dataset)

mod_multi_simple(module_id)
}
\arguments{
\item{id}{shiny id}

\item{dataset}{input dataset}

\item{module_id}{shiny module ID}
}
\description{
This simple module is used for demonstration purposes in documentation, testing and developing
}
\section{Functions}{
\itemize{
\item \code{multi_simple_UI()}: Module UI

\item \code{multi_simple_server()}: Module server

}}
\keyword{internal}
