% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_module.R
\name{mod_odg_dataset_name}
\alias{mod_odg_dataset_name}
\title{A module that exports the current dataset list's name, for testing the output document generation feature}
\usage{
mod_odg_dataset_name(mod_id)
}
\arguments{
\item{mod_id}{shiny module ID}
}
\description{
A module that exports the current dataset list's name, for testing the output document generation feature
}
\keyword{internal}
