% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_module.R
\name{simple_UI}
\alias{simple_UI}
\alias{simple_server}
\alias{mod_simple}
\title{A simple module that counts the number of rows}
\usage{
simple_UI(id)

simple_server(id, dataset)

mod_simple(dataset, module_id)
}
\arguments{
\item{id}{shiny id}

\item{dataset}{input dataset}

\item{module_id}{shiny module ID}
}
\description{
This simple module is used for demonstration purposes in documentation
}
\section{Functions}{
\itemize{
\item \code{simple_UI()}: Module UI

\item \code{simple_server()}: Module server

}}
