% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_module.R
\name{mod_simple2}
\alias{mod_simple2}
\title{A simple module that counts the number of rows}
\usage{
mod_simple2(dataset_name, module_id)
}
\arguments{
\item{module_id}{shiny module ID}
}
\description{
This simple module is used for demonstration purposes in documentation
}
\details{
It is similar to mod_simple but does not use dispatchers
}
