% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_module.R
\name{switch_UI}
\alias{switch_UI}
\alias{switch_server}
\alias{mod_switch}
\title{A module that allows switching to another module}
\usage{
switch_UI(id, name)

switch_server(id, selected, switch_func)

mod_switch(name, selected, switch_func, module_id)
}
\arguments{
\item{id}{the shiny module id+}

\item{name}{A title that will be shown inside the module}

\item{selected}{The name of the tab that we want to switch to}

\item{switch_func}{a function that when passed the name of a tab switches the focus to it}

\item{module_id}{shiny module ID}
}
\description{
This simple module is used for demonstration purposes in documentation
}
\section{Functions}{
\itemize{
\item \code{switch_UI()}: Module UI

\item \code{switch_server()}: Module server

}}
\keyword{internal}
