% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-label.R
\name{name_label_formatter}
\alias{name_label_formatter}
\title{Renames a list to include its value in the name}
\usage{
name_label_formatter(l)
}
\arguments{
\item{l}{a named list}
}
\description{
Renames a list to include its value in the name
}
\details{
If value and name are equal the entry name is not modified.
These lists usually come from swap_names(rpl_nulls_name(get_lbl(dataset)))
}
\keyword{internal}
