% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-label.R
\name{rpl_nulls_name}
\alias{rpl_nulls_name}
\title{Replace NULL values in a list with the name of the entry}
\usage{
rpl_nulls_name(l)
}
\arguments{
\item{l}{a named list.}
}
\value{
A named list.
}
\description{
If one of the values is NULL it will be replaced with the name of the NULL entry.
}
\details{
The input of this functions is usually the output of a \code{get_lbls} call.
}
\keyword{internal}
