% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  data = NULL,
  module_list = list(),
  title = "Untitled",
  filter_data,
  filter_dataset_name = NULL,
  filter_key = if (!is.null(data)) {
     "USUBJID"
 } else {
     NULL
 },
  startup_msg = NULL,
  reload_period = NULL,
  enableBookmarking = "server",
  filter_type = "simple",
  enable_dataset_filter = NULL,
  enable_subgroup = FALSE,
  enable_querychat = FALSE,
  querychat_args = list(),
  filter_default_state = NULL,
  .launch = TRUE,
  .bypass_checks = FALSE,
  .bypass_filter_precomputation = FALSE
)
}
\arguments{
\item{data}{the list of datasets to be used in the Shiny application. This data parameter can be defined as:
\itemize{
\item a named list of lists in which each of the lists will be a list of data.frames. Each of the lists of data.frames
contains a domain, a structure similar to that returned in a \verb{dv.loader::load_data} call.
\item a named list of functions in which each of the functions will return a list of data.frames.
}

All \code{character} variables will be automatically mapped into \code{factors}.}

\item{module_list}{a list of the modules to be included in the Shiny application}

\item{title}{title to be displayed in the browser tab}

\item{filter_data}{( \strong{DEPRECATED} , see \code{filter_dataset_name})}

\item{filter_dataset_name}{a string indicating the name of the dataset used for population filtering.}

\item{filter_key}{a string specifying a common field across all datasets that will be used for population filtering.
Default = "USUBJID" or NULL if no data = NULL}

\item{startup_msg}{a message to be displayed at the start of the application. It can be either NULL or a modal
message defined with shiny::modalDialog.}

\item{reload_period}{Either a lubridate object to specify a duration
or a positive numeric value which is then interpreted as a lubridate duration object in days. By default NULL}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{filter_type}{( \strong{DEPRECATED} )}

\item{enable_dataset_filter}{( \strong{DEPRECATED} )}

\item{enable_subgroup}{A boolean flag indicating if subgroup controls are enabled. The default value is FALSE.}

\item{enable_querychat}{A boolean flag indicating if a chat-based (natural language, powered by the
\code{querychat} package) population filter tab is enabled. Default = FALSE. Requires the \code{querychat},
\code{DBI} and \code{duckdb} packages (and a configured LLM client, see \code{querychat_args}). One chat covers the
whole selected \code{dataset_list}: every dataset in it is written to a shared in-memory database so the
chat can answer cross-dataset questions (e.g. combining \code{filter_dataset_name} with other datasets),
while the resulting subject IDs (from \code{filter_dataset_name}/\code{filter_key}) filter the rest of datasets.}

\item{querychat_args}{A named list of extra arguments forwarded to \code{querychat::QueryChat$new()},
e.g. \code{list(client = "openai/gpt-4.1", greeting = "...", extra_instructions = "...")}. Only used
when \code{enable_querychat = TRUE}.}

\item{filter_default_state}{A JSON string or file (usually exported from the app) that describes the default state of the filter (Only available for \code{development} filters).}

\item{.launch}{by default it should always be TRUE. It should only be false for debugging and testing.
When TRUE it will return the app. When FALSE it will return the options with which the app will be launched.}

\item{.bypass_checks}{by default it should always be FALSE. Only for advanced use. If set to TRUE, the app creator must make sure that
application parameters and modules are correctly configured for all trials loaded in the application, otherwise application may fail. Configuration errors can be checked with a
dry run. To do a dry run use the parameter \code{.launch = FALSE}.}

\item{.bypass_filter_precomputation}{by default it should always be FALSE. Only for advanced use. If set to TRUE, filters are not precomputed per datasetlist.
this increases the app start time and the time required to do a dataset switch.}
}
\description{
This function serves as an entry point for the applications created by the user.
}
