% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-label.R
\name{set_lbl}
\alias{set_lbl}
\title{Set the label attribute of a column in a data frame}
\usage{
set_lbl(df, var, lbl)
}
\arguments{
\item{df}{a dataframe.}

\item{var}{a column of the dataframe}

\item{lbl}{the label}
}
\value{
the dataframe with the replaced label
}
\description{
Set the label attribute of a column in a data frame
}
\keyword{internal}
