% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-label.R
\name{swap_val_names}
\alias{swap_val_names}
\title{Swap names and values}
\usage{
swap_val_names(l)
}
\arguments{
\item{l}{a named vector or list}
}
\value{
A vector or list with swapped names and values
}
\description{
Swap names and values
}
\keyword{internal}
