% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  data = NULL,
  module_list = list(),
  title = "Untitled",
  filter_data = NULL,
  filter_key = if (!is.null(data)) {
     "USUBJID"
 } else {
     NULL
 },
  startup_msg = NULL,
  azure_options = NULL,
  reload_period = NULL,
  enableBookmarking = "server",
  .launch = TRUE
)
}
\arguments{
\item{data}{the list of datasets to be used in the Shiny application. This data parameter can be defined as:
\itemize{
\item a named list of lists in which each of the lists will be a list of data.frames. Each of the lists of data.frames
contains a domain, a structure similar to that returned in a \code{dv.loader::load_data} call.
\item a named list of functions in which each of the functions will return a list of data.frames.
}}

\item{module_list}{a list of the modules to be included in the Shiny application}

\item{title}{title to be displayed in the browser tab}

\item{filter_data}{a string indicating which of the loaded datasets is used for filtering.}

\item{filter_key}{a string specifying a common field across all datasets that will be used to expand the filtering.
Default = "USUBJID" or NULL if no data = NULL}

\item{startup_msg}{a message to be displayed at the start of the application. It can be either NULL or a modal
message defined with shiny::modalDialog.}

\item{azure_options}{A named vector/list containing the options for the Azure Authentication:
Required fields: redirect, resource, tenant, app, version, password.
See AzureAuth::build_authorization_uri for these. If NULL then application will not be secure.
Note that azure_options is only required when the applications handles the authentication by itself. It is not
required in other environments such as POSIT Connect or other environments where the authentication is handled
externally.}

\item{reload_period}{Either a lubridate object to specify a duration
or a positive numeric value which is then interpreted as a lubridate duration object in days. By default NULL}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{.launch}{by default it should always be TRUE. It should only be false for debugging and testing.
When TRUE it will return the app. When FALSE it will return the options with which the app will be launched.}
}
\description{
This function serves as an entry point for the applications created by the user.
}
