% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_group.R
\name{tab_group}
\alias{tab_group}
\title{Create a Tabbed Shiny Module Collection}
\usage{
tab_group(module_list)
}
\arguments{
\item{module_list}{A list of modules.}
}
\value{
The module list marked with an attribute.
}
\description{
\code{tab_group} is a function that groups dv modules in tabsets (see \link[shiny:tabsetPanel]{shiny::tabsetPanel}).
This function is designed to simplify the process of organizing multiple Shiny modules into a single,
visually organized tabset. It allows nested modules.
}
\details{
The function does not make use of \code{namespace} (\code{NS()}) or \code{shiny::moduleServer} to implement traditional Shiny
modules, but creates a tab-like UI structure, making it an aesthetic modification.

Outputs of grouped modules are accessible by using the module id, there is no special change required.

When switching to grouped outputs the id of the module can be used directly. This approach is incompatible
with the \code{switch2} function and must be used with \code{switch2mod}.
}
